Drop useCommandItem hook and simplify CommandUniverseGridItem
The grid is browse-only, so CommandUniverseGridItem doesn't need
drag-and-drop, selection, scroll-into-view, lottie animation, or
gesture-progress highlighting. Inline the small amount of logic
CommandTableItem actually uses (drag source, scroll-into-view,
label/description resolution) directly into that component, and
delete the shared hook.
